+++ src/cli/commands/spectrum-analysis.ts
@@ -14,7 +14,7 @@ import {
     serviceCredentialLog,
     verboseLog,
 } from "./command-utils";
-import ora = require("ora");
+import ora = require("ora-classic");
 const mime = require("mime-types");
 let color = getColor("blue");