+++ templates/ts/ts-template.ts
@@ -11,7 +11,7 @@ export const tstemplate: string = `import { DataPointValue, MindConnectAgent, Mi
 
     // onboarding the agent
     // Check in the local agent state storage if agent is onboarded.
-    // https://opensource.mindsphere.io/docs/mindconnect-nodejs/agent-development/agent-state-storage.html
+    // https://developer.siemens.com/industrial-iot-open-source/mindconnect-nodejs/agent-development/agent-state-storage.html
     if (!agent.IsOnBoarded()) {
         // wrapping the call in the retry function makes the agent a bit more resillient
         // if you don't want to retry the operations you can always just call await agent.OnBoard(); instead.